+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 # 0.12.0 
 - If model is unable to properly compute metrics, it's value is replaced with np.nan
-- Added parameter in split method to encode categorical columns
+- Added 'encode' parameter in split method to encode categorical columns
+- Added 'missing' values parameter in split method for filling missing values for both numerical and categorical columns.
 # 0.11.21 - BUG FIX
 - Added missing requirement for kaleido engine
 # 0.11.0 - PATCH RELEASE